Now Kitspace shows PCB correctly
I could not figure out why the PCB was yellow. All zones were filled with yellow and the silk text was not visible. Looked horrible. However, other viewers worked correctly (used oshpark.com viewer). The problem was that this viewer is using all possible layers to render the view. And if one is missing, it starts to show buggy image(s)