Update IR sensor to v0.4
IR sensor board updated to version 0.4:
- removed MOSFET switch;
First of all, it was supposed to be IRFL44 , L version which can be used with logic levels. But instead we put IRFZ which barely opens at 3.3 V and thus cannot be effectively used here. Moreover, during the actual usage of the board, 5V line was constantly ON and it did not cause to much of a trouble. This change is also correlated with the upcoming redesign of the main hub.
- Add additional phototransistor connector
- Add additional IR LED connector

Co-Authored-By: Andre Maia Chagas <a.maia-chagas@sussex.ac.uk>